requirements.txt
black==22.8.0
click==8.1.3
colorama==0.4.5
fastapi==0.68.0
h11==0.13.0
mypy-extensions==0.4.3
numpy==1.23.3
opencv-python==4.6.0.66
pathspec==0.10.1
Pillow==9.2.0
platformdirs==2.5.2
pydantic==1.10.2
python-dotenv==0.21.0
starlette==0.14.2
tomli==2.0.1
typing_extensions==4.3.0
uvicorn==0.12.2
